Fixing return values for the interface because suddenly everyone else decided that it was a better idea to be double instead of int

+++ src/app/Calc.java
@@ -6,5 +6,5 @@ package app;
  */
 public interface Calc {
 
-    int operate(String input);
+    double operate(String input);
 }
